feat(config): add ability for users to also use the server.cfg
Will allow users the option to use either the config.json or server.cfg. This is a change I'm mostly making for myself, but am opening the pr in case its something that may have been of interest to others. These changes are not breaking to the current configuration specifications.

The README has been updated accordingly. 

in `server.lua` the load/ready events have been changed to allow for more consistent (100% consistent) startup logic.

I would have just liked to use a mysql connection string for the server.cfg, but maybe at a later date.